requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
mysql-connector-python==8.3.0
requests
urllib3==1.26.10
